Add update profile in account modal
